fix: exhaustively cover all activity statuses
When mapping from AniList activity to Discord embed, we took care of a
few common `status` forms but left out the others to depend on a default
format:

  {} {} of [{}]({})

This gives us 'plans to watch ? of [title](url)' for 'planning'
activities.

Lets define mappings for all the statuses known to us explicitly.