refactor: rid unused image fields
Since we have no use of 'banner image' AniList gives us, let's get rid
of it from our model. We also did not have it in our GraphQL query, but
still for some reason had it in our model. The field was therefore
always None.

On the Discord side, we use thumbnail to show the cover images of media
and have no use of the image field.

Let's get rid of these.